Initial commit
Initial commit, far from complete.

One major issue that needs to be sorted before this is even usable is
time-zones.

The NOAA returns the forecast in _either_ UTC or local time, and there's
no way to programatically figure out what timezone local time is. If we
choose to query in UTC, we get results for today's high- and low-tides
within the bounds of midnight-to-midnight UTC. For Pacific time, that's
like 5pm-5pm, which is very awkward.