Fix the timezone issues
Turns out elisp has enough date/time facilities to make this relatively
simple.

If requests are made asking for GMT data, we can present the information
in the local timezone. Displaying a time in UTC is dead simple.

The actual UI here requires a pretty big overhaul now that this is
possible, but at least it's useful again.